Rename comparasion{,.test}.js files to merge-compare.{,.test}.js
