+++ src/profile-logic/merge-compare.js
