+++ src/test/fixtures/profiles/processed-profile.js
@@ -12,7 +12,7 @@ import {
   getEmptyUnbalancedNativeAllocationsTable,
   getEmptyBalancedNativeAllocationsTable,
 } from '../../../profile-logic/data-structures';
-import { mergeProfilesForDiffing } from '../../../profile-logic/comparison';
+import { mergeProfilesForDiffing } from '../../../profile-logic/merge-compare';
 import { stateFromLocation } from '../../../app-logic/url-handling';
 import { UniqueStringArray } from '../../../utils/unique-string-array';
 import { ensureExists } from '../../../utils/flow';
